requirements.txt
pycryptodomex==3.15.0
requests==2.28.1
